requirements.txt
ipyleaflet>=0.14.0
ipywidgets>=7.7.1
matplotlib>=3.7.0
numpy>=1.24.2
opencv-python>=4.1.2
pandas>=1.5.3

flask 
python-dotenv>=0.20.0
requests>=2.27.1
scikit-learn
seaborn>=0.12.2
streamli
sympy>=1.10.1
gunicorn